+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## [1.0.0] - 14/11/2021.
 
-* Added makeMove with SAN notation 
+* Added full null safety
+* Major refactor
+* Added new controller architecture
+* Removed scoped_model
+* Added capability to add arrows